grammar: Prioritize choice of name type
It's surprising that when LARK was given a choice, it did not seem to
pick the most specific type.  In other words something like

FIDL_NAME: FIDL_DEREFENCED_NAME | FIDL_FQ_NAME | FIDL_NQ_NAME

... even when given a non-qualified name it ends up being a FIDL_NAME
token instead of FIDL_NQ_NAME.  Not sure if it's some other
misunderstanding, and also not sure if this is the best way, but now it
is forced to prefer NQ and then FQ and only as last resort FIDL_NAME.

Now that I think about it, replacing FIDL_NAME with a token that
is hidden in the tree seems to be the right solution:

?fidl_name: FIDL_DEREFENCED_NAME | FIDL_FQ_NAME | FIDL_NQ_NAME

... this might be the right approach but this requires changes in the
processing of the resulting tree, which is too much at the moment.